+++ slh_sh/main.py
@@ -7,6 +7,8 @@ import slh_sh.commands.go as go
 import slh_sh.commands.sync as sync
 import slh_sh.commands.self as self
 
+from slh_sh.commands.query import query
+
 app = typer.Typer()
 
 app.add_typer(extract.app, name="extract")
@@ -15,6 +17,8 @@ app.add_typer(add.app, name="add")
 app.add_typer(go.app, name="go")
 app.add_typer(sync.app, name="sync")
 app.add_typer(self.app, name="self")
+
+app.command()(query)
 
 if __name__ == "__main__":
     app()